Move internal/ecs/events/event.go → internal/events/event.go.
