+++ internal/ecs/systems/score_system.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/juanancid/maze-adventure/internal/ecs"
 	"github.com/juanancid/maze-adventure/internal/ecs/components"
-	"github.com/juanancid/maze-adventure/internal/ecs/events"
+	"github.com/juanancid/maze-adventure/internal/events"
 )
 
 type ScoreSystem struct{}
