+++ internal/ecs/world.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 
 	"github.com/juanancid/maze-adventure/internal/ecs/components"
-	"github.com/juanancid/maze-adventure/internal/ecs/events"
+	"github.com/juanancid/maze-adventure/internal/events"
 )
 
 type World struct {
