+++ internal/events/event.go
